### 2021-02-05 #3014 @kyleniemeyer
Hi @marcozennaro, thanks for your interest in JOSS. We use this pre-review issue to assign the editor and reviewers, and also address any obvious issues.

First, your software repository is missing a valid open-source license. Please select a license, and add it.

Also, in your paper, I see some section/subsection references that do not appear to be compiling correctly—please take a look.
