+++ paper.bib
@@ -12,7 +12,7 @@
   author={Zennaro, Marco and Pietrosemoli, Ermanno and Rainone, Marco and Trinchero, Daniele and Poletti, Mattia and Colucci, Giovanni},
   booktitle={Proceedings of the 1st Workshop on Experiences with the Design and Implementation of Frugal Smart Objects},
   pages={24--29},
-  doi={https://doi.org/10.1145/3410670.3410856},
+  doi={10.1145/3410670.3410856},
   year={2020}
 }
 
