### 2021-02-26 #3062 @gkthiruvathukal
Thanks, @kthyng! Possibly some sort of concurrency/race condition. I was pretty sure everything was ok, since the paper built just fine prior to going to REVIEW. Perhaps we can have @arfon take a look at this (happy to record an issue in GitHub but not quite sure where to do it).
