### 2021-03-17 STI-ICTP/TropPOLoRaTools#1 @wtbarnes
Make TropPOLoRaTools an installable package
Currently, this piece of software represents a useful collection of scripts. However, in order to better facilitate ease of use by outside users and enable easier workflows, it may be useful to make this into an installable library. See [this guide](https://python-packaging-user-guide.readthedocs.io/tutorials/packaging-projects/) for a walkthrough of how to create installable Python packages.
