### 2021-03-17 STI-ICTP/TropPOLoRaTools#2 @wtbarnes
Add documentation
Currently, the only documentation for this software lies in PDF format under `docs`. It would be preferable if the documentation were instead formatted in a markup language and rendered as a webpage (e.g. using tools like [Sphinx](https://www.sphinx-doc.org/en/master/) or [mkdocs](https://www.mkdocs.org/)). This would allow the documentation to be version controlled and thus more easily maintained. In this format, the documentation could also be hosted on a service such as [Read the Docs](https://readthedocs.org/) that would allow it to be more discoverable.

Additionally, tools like Sphinx would also support the automatic generation of API documentation that could be included in the built documentation.
