### 2021-03-17 STI-ICTP/TropPOLoRaTools#1 @wtbarnes
This would also greatly ease the problem of installing all needed dependencies as these could be listed in the `setup.py` (or `setup.cfg`) file and the installation handled by `pip`.
