### 2021-03-17 STI-ICTP/TropPOLoRaTools#3 @wtbarnes
Add contribution guidelines and code of conduct
To enable users and community members to more easily contribute to this software, a guide for contributing should be added to the repository. See [this page](https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors#examples-of-contribution-guidelines) for more info. It is also recommended that repositories have a code of conduct. See [this page](https://docs.github.com/en/github/building-a-strong-community/adding-a-code-of-conduct-to-your-project) for more details.
