### 2021-03-17 STI-ICTP/TropPOLoRaTools#4 @wtbarnes
Add testing
As far as I can tell, TopPOLoRaTools does not contain any automated or manual tests. The [pytest](https://docs.pytest.org/en/stable/) testing framework would likely be useful in implementing a set of both unit and integration tests.

It may be helpful to first address #1 before adding a test suite.

Additionally, once a testing framework is implemented, it would be useful to run these tests automatically through a continuous integration service like [GitHub Actions](https://docs.github.com/en/actions) to ensure that the tests continue to pass, in a platform-independent manner, as the code base evolves.
