### 2021-03-17 #3062 @wtbarnes
@marcozennaro @gkthiruvathukal I've finished my review and apologies for the delay. Unfortunately, at this stage, I cannot recommend this paper for publication in JOSS. I have completed the reviewer check list above and am currently forced to leave many of the boxes unchecked as the software does not meet these requirements. Without addressing the fundamental issues detailed below, it will be difficult to proceed with this review. However, I hope that the feedback from this review will be helpful in getting the code and paper into a publishable state and that the paper can be accepted.

Below, I have outlined several fundamental and minor issues. In order for the paper and code to be published, I believe that the fundamental issues should be addressed. The minor issues are just suggestions.

### Fundamental Issues

* **Authorship:** From the commit history on GitHub, it appears that while the primary author wrote the paper and corrected various DOI and licensing errors, the actual software was written by another user, presumably the second author on this paper. It appears that the git history for this package is relatively short and largely confined to one or two commits so if this git history is misleading, this should be clarified and this comment can be disregarded. As it stands, this is in conflict with checklist item number 3 under "General Checks" regarding contribution and authorship.
* **Packaging/Installation:** There is no clear way to install this software, including installing all of the required dependencies. I strongly recommend that the authors restructure this software as a Python package and have provided more details in the GitHub issue listed below. At the very least, the authors should provide a `requirements.txt` file such that users can easily install all required dependencies with a single command.
* **Documentation:** While the PDF in the `doc` directory does contain some instructions for running the software, there is no formal API documentation. Furthermore, while there is a list of package dependencies, there is no clearly described method for easily installing these dependencies or a delineation between which dependencies are part of the Python standard library and which are third-party dependencies.
* **Testing:** There appear to be no automated tests included with the software, either unit tests or integration tests. I highly recommend that the authors consider using a common testing framework (e.g. [pytest](https://docs.pytest.org/en/stable/)) to implement a test suite for their software.
* **Paper:** While the paper does contain a Statement of Need section, it is not clear who the audience for the software is or how it compares to existing tools for accomplishing the same tasks. Furthermore, I believe that the explanation of the scope of the problem could be at a bit higher level and ideally with less domain-specific language (e.g. "radiosonde" is never clearly defined). I believe the "Use Cases" section could also be removed from the text of the paper and be formatted as a set of useful examples in the software documentation. There are also a number of typographical and formatting errors that need to be corrected.
* **Community:** There is no clear set of guidelines for how to contribute to the package or a code of conduct.

### Minor Issues

* The naming of the package is inconsistent between the paper submission and the repository. Is the software called `TAT.py` or `TropPOLoRaTools`?
* The use of Google Colab exclusively for the example notebook creates a barrier for new users by forcing them to have a Google account in order to run the software. While Colab is an extremely useful tool, I would recommend disentangling the notebook itself from running on Colab and instead provide a link in the README that automatically launches this on Google Colab or [mybinder](https://mybinder.org/).
* The code does not appear to follow any standard style guide. I would recommend consulting [PEP-8](https://www.python.org/dev/peps/pep-0008/). These types of style guides can be applied automatically using a number of formatting tools (e.g. [`flake8`](https://flake8.pycqa.org/en/latest/#)).
* There appears to be a large amount of duplicated code between each `.py` file. I would strongly recommend the authors pull out commonly-used functions and place them in a single file and import those functions/classes from a single source rather than repeating them across multiple files. I believe that a reorganization of the code into a Python package with standard structure would greatly solve this issue. This would also make the software much more easy to test and thoroughly document.
* Binary files such as PDFs and even Jupyter notebooks are difficult to version control and can often greatly increase the size of the repository. I would recommend removing these large binary files and instead generate them on the fly (e.g. using tools like [Sphinx](https://www.sphinx-doc.org/en/master/))
* There is not currently any information in the README of the repository about how to run or install the software. I would recommend that the authors either provide a very minimal guide to getting started in the README or a link to the documentation.

### Repository Issues

I've created the following issues on the `TropPOLoRaTools` repository to address the points raised in the sections above

* STI-ICTP/TropPOLoRaTools#1
* STI-ICTP/TropPOLoRaTools#2
* STI-ICTP/TropPOLoRaTools#3
* STI-ICTP/TropPOLoRaTools#4
