### 2021-03-18 #3062 @wtbarnes
> I would like immediate clarification on the authorship issue raised by @wtbarnes. Please note that the issues raised do not disqualify you or your co-authors, but we'd like to get a more clear statement of how you've contributed to this effort. It is my view that contributions are in no way limited to coding, but if your co-author is the primary contributor, I think there is a case to list that person first.

My comment about authorship was based on commits to the master branch: https://github.com/STI-ICTP/TropPOLoRaTools/commits/master. The git history for this project is relatively short so I of course understand that much work could have been done on the software prior to the beginning of the commit history and that much information could very well be lost due to the changes to the software being confined to the first two commits. I also fully understand that git commits are a very imperfect measure of contribution to a project and fully agree that contributions are not limited to code contributions. I only mean to ask for additional clarification on this from the primary author.
