### 2021-09-30 #3062 @danielskatz
Given that it's been over 6 months since the author has responded, I suggest that if there is no progress in the next week, we reject the paper and allow the author to resubmit at a future time.
